requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
pillow==10.3.0
pyTelegramBotAPI==4.17.0
requests==2.31.0
telebot==0.0.5
urllib3==2.2.1
